refactor(repo): extract the @intake/core package (phase 3b)
Second half of Phase 3: a new pure, isomorphic @intake/core package
(depends only on @intake/types) holding the cleanly I/O-free domain logic,
consumed as raw TS via transpilePackages. Re-export shims at the original
@/lib/* and @/hooks/* paths keep every importer unchanged (zero churn).

Moved whole: alcohol-units, compound-utils (its CompoundStrength import
repointed @/lib/db -> @intake/types/records), progress-utils, settings-helpers.

Partial splits (pure subset -> core, browser/impure parts stay in apps/web):
- security: the 4 DOM-free fns (sanitizeNumericInput, sanitizeTextInput,
  sanitizeForAI, sanitizeReportText + internal redactPii) -> @intake/core/security;
  obfuscate/deobfuscate (btoa/atob) + isSecureContext/getSecurityWarnings stay.
- use-shake-gesture: the pure createShakeDetector state machine (+ ShakeSample,
  ShakeDetectorConfig) -> @intake/core/shake; the React hook + permission fns stay.
- service-result: the ServiceResult/PaginatedResult types -> @intake/types/service;
  the ok/err/unwrap helpers -> @intake/core/service.

Purity guard: @intake/core's tsconfig ships no DOM lib and no @types/node, so
window/document/localStorage/btoa/atob are compile errors — a structural,
zero-config purity guard (which is why the btoa/atob obfuscate fns stayed in
the app). Deferred to Phase 3.1 / later (see DEFERRED.md): the wall-clock
now/tz injection of the 2 analytics fns, migrateSettings (entangled with app
constants + the Settings store type), the btoa/atob fns, the dedicated
purity-lint rule, and the eventual importer rewrites off the @/lib/* shims.

Phase 3b of the Turborepo migration; completes the types+core layering spine.

Verified: turbo typecheck (4 pkgs, @intake/core under the DOM-free guard) +
lint (0 errors), 2017 unit tests, 33 integration (testcontainers), production
build (transpilePackages JIT + client-only boundary holds), bundle-security
scan, re-export completeness across all 7 moved modules, 3-agent adversarial
review (CI / boundary / regression) — no blockers.
