+++ packages/core/src/service-result.ts
@@ -1,0 +1,14 @@
+import type { ServiceResult } from "@intake/types/service";
+
+export function ok<T>(data: T): ServiceResult<T> {
+  return { success: true, data };
+}
+
+export function err<T = never>(error: string, details?: unknown): ServiceResult<T> {
+  return { success: false, error, details };
+}
+
+export function unwrap<T>(result: ServiceResult<T>): T {
+  if (result.success) return result.data;
+  throw new Error(result.error);
+}
